requirement.txt
asgiref==3.8.1
Django==5.0.6
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.2
dnspython==2.6.1
drf-yasg==1.21.7
inflection==0.5.1
mongoengine==0.28.2
packaging==24.1
PyJWT==2.8.0
pymongo==4.7.3
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
